set size_t

+++ string.cpp
@@ -33,15 +33,15 @@ string::string(const string& s){
 	}	
 }
 //getters------------------------------------------------------------
-int string::capacity(){
+size_t string::capacity(){
   return capacity_;
 }
 
-int string::length(){
+size_t string::length(){
 	return len_;
 }
 
-int string::max_size(){
+size_t string::max_size(){
 	return max_size_;
 }
 //destructor--------------------------------------------------------------
@@ -59,7 +59,7 @@ char* string::c_str(){
 	return c;
 }
 
-int string::size(){
+size_t string::size(){
 	return len_;
 }
